--- include/ion-bb-dnn/test.cc
@@ -1,5 +1,0 @@
-#include "yolov4_tfl_rt.h"
-int main(int argc, char *argv[])
-{
-    return func(argc, argv);
-}
